规则引擎量化人群参与活动

+++ Lottery-ERP/src/main/java/com/banana69/lotteryERP/interfaces/sys/mapper/UserMapper.java
@@ -16,5 +16,5 @@ import java.util.List;
  */
 public interface UserMapper extends BaseMapper<User> {
 
-    public List<String> getRoleNameByUserId(Integer userId);
+    public List<String> getRoleNameByUserId(Long userId);
 }
